Patch console in CLI and workers
