+++ packages/core/logger/src/Logger.js
@@ -3,6 +3,7 @@
 import type {IDisposable, LogEvent} from '@parcel/types';
 
 import {ValueEmitter} from '@parcel/events';
+import {inspect} from 'util';
 
 class Logger {
   #logEmitter = new ValueEmitter<LogEvent>();
@@ -58,3 +59,44 @@ class Logger {
 
 const logger = new Logger();
 export default logger;
+
+let consolePatched = false;
+
+// Patch `console` APIs within workers to forward their messages to the Logger
+// at the appropriate levels.
+// TODO: Implement the rest of the console api as needed.
+// TODO: Does this need to be disposable/reversible?
+export function patchConsole() {
+  if (consolePatched) {
+    return;
+  }
+
+  /* eslint-disable no-console */
+  // $FlowFixMe
+  console.log = console.info = (...messages: Array<mixed>) => {
+    logger.info(joinLogMessages(messages));
+  };
+
+  // $FlowFixMe
+  console.debug = (...messages: Array<mixed>) => {
+    // TODO: dedicated debug level?
+    logger.verbose(joinLogMessages(messages));
+  };
+
+  // $FlowFixMe
+  console.warn = (...messages: Array<mixed>) => {
+    logger.warn(joinLogMessages(messages));
+  };
+
+  // $FlowFixMe
+  console.error = (...messages: Array<mixed>) => {
+    logger.error(joinLogMessages(messages));
+  };
+
+  /* eslint-enable no-console */
+  consolePatched = true;
+}
+
+function joinLogMessages(messages: Array<mixed>): string {
+  return messages.map(m => (typeof m === 'string' ? m : inspect(m))).join(' ');
+}
