+++ packages/core/parcel/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@parcel/config-default": "^2.0.0-alpha.0",
     "@parcel/core": "^2.0.0-alpha.0",
+    "@parcel/logger": "^2.0.0-alpha.0",
     "chalk": "^2.1.0",
     "commander": "^2.19.0",
     "get-port": "^4.2.0",
